Issue 29 : allow domain events to be processed within or outside of the main transaction
The use of marker interface now allows the flexibility for the domain event handlers to be processed inside or outside the main transaction.
For a DomainEventHandler to be executed outside of the main transaction, one only needs to tag it with the marker interface ToProcessAfterMainTransaction.

To illustrate this, the ItemBorrowedDomainEventHander is used as an exemple in the Spring boot project